+++ js/data.js
@@ -1,5 +1,6 @@
-import {getRandomArrayElemnt} from './util.js';
+import {getRandomArrayElement} from './util.js';
 import {getRandomInteger} from './util.js';
+import {createId} from './util.js';
 
 const PICTURE_COUNT = 25;
 const AVATAR_COUNT = 6;
@@ -28,32 +29,24 @@ const NAMES = [
   'Вашингтон',
 ];
 
-const createId = () => {
-  let lastId = 0;
-  return () => {
-    lastId += 1;
-    return lastId;
-  };
-};
-
 const commentId = createId();
 
 const createMessage = () => Array.from(
   {length:getRandomInteger(1,2)},
-  () => getRandomArrayElemnt(COMMENT_LINES),
+  () => getRandomArrayElement(COMMENT_LINES),
 ).join(' ');
 
 const createComment = () => ({
   id: commentId(),
   avatar: `img/avatar-${getRandomInteger(1, AVATAR_COUNT)}.svg`,
   massege: createMessage(),
-  name: getRandomArrayElemnt(NAMES),
+  name: getRandomArrayElement(NAMES),
 });
 
 const createPicture = (index) => ({
   id: index,
   url: `photos/${index}.jpg`,
-  description: getRandomArrayElemnt(DESCRIPTION),
+  description: getRandomArrayElement(DESCRIPTION),
   likes: getRandomInteger(LIKE_MIN, LIKE_MAX),
   comment: Array.from(
     {length: getRandomInteger(0, COMMENT_COUNT)},
