+++ js/util.js
@@ -5,7 +5,14 @@ const getRandomInteger = (a, b) => {
   return Math.floor(result);
 };
 
-const getRandomArrayElemnt = (items) => items[getRandomInteger(0, items.length - 1)];
+const getRandomArrayElement = (items) => items[getRandomInteger(0, items.length - 1)];
 
-export {getRandomArrayElemnt};
-export {getRandomInteger};
+const createId = () => {
+  let lastId = 0;
+  return () => {
+    lastId += 1;
+    return lastId;
+  };
+};
+
+export {getRandomArrayElement, getRandomInteger, createId};
